ops/dev-stack/datagen/src/requirements.txt
# Database connectivity
psycopg2-binary==2.9.9

# Fake data generation
Faker==22.0.0
